Poisson regression models

poisson_reg() defines a generalized linear model for count data that follow a Poisson distribution. There are different ways to fit this model. See the engine-specific pages for more details:

glm (default)

hurdle

zeroinfl

glmnet

stan

More information on how parsnip is used for modeling is at https://www.tidymodels.org/.

poisson_reg(
  mode = "regression",
  penalty = NULL,
  mixture = NULL,
  engine = "glm"
)

# S3 method for poisson_reg
update(
  object,
  parameters = NULL,
  penalty = NULL,
  mixture = NULL,
  fresh = FALSE,
  ...
)

Arguments

mode
A single character string for the type of model. The only possible value for this model is "regression".

penalty
A non-negative number representing the total amount of regularization (glmnet only).

mixture
A number between zero and one (inclusive) that is the proportion of L1 regularization (i.e. lasso) in the model. When mixture = 1, it is a pure lasso model while mixture = 0 indicates that ridge regression is being used. (glmnet and spark only).

engine
A single character string specifying what computational engine to use for fitting.

object
A boosted tree model specification.

parameters
A 1-row tibble or named list with main parameters to update. If the individual arguments are used, these will supersede the values in parameters. Also, using engine arguments in this object will result in an error.

fresh
A logical for whether the arguments should be modified in-place of or replaced wholesale.

...
Not used for update().

Value

An updated model specification.

Details

This function only defines what type of model is being fit. Once an engine is specified, the method to fit the model is also defined.

The model is not trained or fit until the fit.model_spec() function is used with the data.

References

https://www.tidymodels.org, Tidy Models with R

See also

Examples

poisson_reg()
#> Poisson Regression Model Specification (regression)
#> 
#> Computational engine: glm 
#> 

# Model from Agresti (2007) Table 7.6
log_lin_mod <-
  poisson_reg() %>%
  set_engine("glm") %>%
  fit(count ~ (.)^2, data = seniors)

summary(log_lin_mod$fit)
#> 
#> Call:
#> stats::glm(formula = count ~ (.)^2, family = stats::poisson, 
#>     data = data)
#> 
#> Deviance Residuals: 
#>        1         2         3         4         5         6         7         8  
#>  0.02044  -0.02658  -0.09256   0.02890  -0.33428   0.09452   0.49134  -0.03690  
#> 
#> Coefficients:
#>                           Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)                5.63342    0.05970  94.361  < 2e-16 ***
#> marijuanayes              -5.30904    0.47520 -11.172  < 2e-16 ***
#> cigaretteyes              -1.88667    0.16270 -11.596  < 2e-16 ***
#> alcoholyes                 0.48772    0.07577   6.437 1.22e-10 ***
#> marijuanayes:cigaretteyes  2.84789    0.16384  17.382  < 2e-16 ***
#> marijuanayes:alcoholyes    2.98601    0.46468   6.426 1.31e-10 ***
#> cigaretteyes:alcoholyes    2.05453    0.17406  11.803  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for poisson family taken to be 1)
#> 
#>     Null deviance: 2851.46098  on 7  degrees of freedom
#> Residual deviance:    0.37399  on 1  degrees of freedom
#> AIC: 63.417
#> 
#> Number of Fisher Scoring iterations: 4
#> 

# ------------------------------------------------------------------------------

library(pscl)
#> Classes and Methods for R developed in the
#> Political Science Computational Laboratory
#> Department of Political Science
#> Stanford University
#> Simon Jackman
#> hurdle and zeroinfl functions by Achim Zeileis

data("bioChemists", package = "pscl")

poisson_reg() %>%
  set_engine("hurdle") %>%
# Extended formula:
  fit(art ~ . | phd, data = bioChemists)
#> parsnip model object
#> 
#> Fit time:  35ms 
#> 
#> Call:
#> pscl::hurdle(formula = art ~ . | phd, data = data)
#> 
#> Count model coefficients (truncated poisson with log link):
#> (Intercept)     femWomen   marMarried         kid5          phd         ment  
#>     0.67114     -0.22858      0.09648     -0.14219     -0.01273      0.01875  
#> 
#> Zero hurdle model coefficients (binomial with logit link):
#> (Intercept)          phd  
#>      0.3075       0.1750  
#> 

model <- poisson_reg(penalty = 10, mixture = 0.1)
model
#> Poisson Regression Model Specification (regression)
#> 
#> Main Arguments:
#>   penalty = 10
#>   mixture = 0.1
#> 
#> Computational engine: glm 
#> 
update(model, penalty = 1)
#> Poisson Regression Model Specification (regression)
#> 
#> Main Arguments:
#>   penalty = 1
#>   mixture = 0.1
#> 
#> Computational engine: glm 
#> 
update(model, penalty = 1, fresh = TRUE)
#> Poisson Regression Model Specification (regression)
#> 
#> Main Arguments:
#>   penalty = 1
#> 
#> Computational engine: glm 
#>